saving my work

+++ models/stockOrder.js
@@ -10,9 +10,8 @@ module.exports = function (sequelize, Sequelize) {
       type: Sequelize.INTEGER
     },
     purchasePrice: {
-      type: Sequelize.INTEGER
+      type: Sequelize.FLOAT(2)
     }
   })
-
   return StockOrder
 }
